requirements.txt
flask
pandas
scikit-learn
joblib
flask-cors
PyJWT
fpdf
pypdf
gunicorn